% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReproducibleCDI-package.R
\docType{package}
\name{ReproducibleCDI-package}
\alias{ReproducibleCDI}
\alias{ReproducibleCDI-package}
\title{ReproducibleCDI: Reproducing the Community Deprivation Index of Neighborhood Socioeconomic Status}
\description{
Provides a simple, tidy interface for calculating the Community Deprivation Index (CDI), a composite measure of neighborhood-level socioeconomic deprivation, using data from the US Census Bureau's American Community Survey (ACS). The CDI follows the Singh/Area Deprivation Index (ADI) methodology - a principal-component-weighted composite of 18 standardized ACS variables computed natively at the census block group level - and adds a percentage-uninsured variable. Features include automatic data fetching, flexible geographies (county, tract, block group), principal component analysis, hierarchical (block group -> tract -> county) imputation, and rescaling to a 1-100 percentile ranking.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HanLabCollaboration/ReproducibleCDI}
  \item Report bugs at \url{https://github.com/HanLabCollaboration/ReproducibleCDI/issues}
}

}
\author{
\strong{Maintainer}: Rafael Catoia Pulgrossi \email{rcatoiap@stanford.edu}

Authors:
\itemize{
  \item Tae Yoon Lee \email{harrytyl@stanford.edu}
}

}
\keyword{internal}
